print list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
   */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int node_count = 0;
+	size_t node_count = 0;
 
 	while (h != 0)
 	{
